ENH: Rename module from `Tracking3D` to `Hierarchical3DRegistration`
