+++ CMakeLists.txt
@@ -44,7 +44,7 @@ add_subdirectory(AutoscoperM)
 add_subdirectory(TrackingEvaluation)
 add_subdirectory(CalculateDataIntensityDensity)
 add_subdirectory(VirtualRadiographGeneration)
-add_subdirectory(Tracking3D)
+add_subdirectory(Hierarchical3DRegistration)
 ## NEXT_MODULE
 
 #-----------------------------------------------------------------------------
