+++ Hierarchical3DRegistration/CMakeLists.txt
@@ -1,5 +1,5 @@
 #-----------------------------------------------------------------------------
-set(MODULE_NAME Tracking3D)
+set(MODULE_NAME Hierarchical3DRegistration)
 
 #-----------------------------------------------------------------------------
 set(MODULE_PYTHON_SCRIPTS
